[Standard Deviation Console Application - Refactoring + Error handling]
Refactored the code to have separate classes and functions to hold and perform the different calculations.
A stat class was added to perform the mean, variance, and standard deviation.
A data file reader class was added to perform the parsing of the data given to it.

Error handling was added for the following:
-parsing is attempted first, and if it fails, the data reader will skip over the failed data parse element
-Argument count is checked to make sure an argument is actually given
-File not found exception is caught and an error is printed to let user know the file doesn't exist
-Variance division makes sure a division by 0 is not allowed, either being 1 or Count - 1 if there is more than 1 element