[Calculator Console Application - Interpreter Pattern]
Created a console application that uses the interpreter pattern to evaluate a mathematical expression.
There is no error handling at the moment and it is a rudementary implementation but it is able to correctly evaluate:
-Operation Order
-Signed values

The code needs some refactoring but I want to think over it a bit more to see if there is a way to make an operation tree that can be executed or if there is a more intuitive way of evaluating the expressions.
Currently the string is filtered, then ran through an expressions interface that is used to evaluate itself, and then converted into a string, and ran again for evaluation of remaining operations. So I'd like to refactor some of it later after sitting down and writing a bit of psuedocode to see if there is a cleaner way of approaching the problem and making the interpreter pattern implemented better in this example.