[Calculator Console Application - Syntax and Token Generator]
Code is being refactored to better follow the interpreter patter design.
Now the source string argument is read through and tokens are generated to separate the expression into it's syntactical pieces.
The Token Generator is not in charge of handling syntax validation, it does currently however validate only the first token. I would like to refactor this away eventually so it is not required in order to function correctly.
The parser validates the syntax of the tokens first before constructing it's expression tree.

Error handling and exceptions were introduced now to catch invalid inputs or expressions.
Currently the application can identify tokens, create them, and validate expression syntax for arithmetic.