[Calculator Console Application - Expression Tree Construction]
Parser can now take all tokens, and construct an expression tree which can be used to evaluate the parsed expression.
It does so recursively, starting from the right most side of the expression, and working it's way leftward to the beginning of the expression.
This way when it is evaluated, it will evaluate bottom up, where the bottom holds the left most side of the arithmetic equation and the top holds right most side.

Currently is working and valid for addition and subtraction operations, it can also evaluate multiplication and division however it is done so in the wrong order.
Next revision should have the multiplication and division solved.