[Calculator Console Application - Correct Arithmetic Expression Tree]
The interpreter pattern is now completely implemented. The parser can now build the expression tree for the given arithmetic expression supplied by the user for +, -, /, *, %.
It does so in a recursive pattern, reading from right to left, and when hitting a multiplication, division, or modulus operator it will branch off, starting a recursion to the left from the left most addition or subtraction operator, and then start a recursion to the right reading all the multiplcation, division, and modulus operators until it hits the end or an unnecessary operator.

Error and exception handling has been added to the arithmetic evaluations to check for arithmetic overflow and division by zero.
The parser class has had some of the recursive functions refactored into separate functions to help with reading the code.
However I would like to go through a final refactoring stage to hopefully move the construction of the arithmetic expression to it's own class.