[Calculator Console Application - Refactoring + Modulus Error]
Refactored the code to try and make each class self contained in what it does.
Reduced line numbers and length wherever possible.

Found an error in operation order for the multiplication, division, and modulus operations.
While Multiplication and division are commutative, the modulus operation is not, and this has caused an issue with the direction the expressions are built.
Due to this the ExpressionBuilder will have to be changed now to build the multiplication, division, and modulus operations from right to left, instead of from left to right.