[Calculator Console Application - Modulus Correction + Refactoring]
So the modulus error was solved by basically making it to instead of having multiply, divide, and modulus operations be solved right to left, they are now solved left to right.
To do this, the code needed to be refactored for clarity as now it does not move the reading position ahead and then start reading the operations from left to right.
Instead it will read ahead and find the next add or subtract operation, then group the multiply and divide operations as the right operand of the add or subtract operation found, or if none is found simply return the grouped operations.

The functions responsible for this were renamed to readAhead and parsePriorityExpressions, as that better describes what it is doing when it runs into a multiply, divide, or modulus operator.