[References and Null - Experiment]
Experimented with ways of being able to pass linked references which if 1 were set to null the rest would be too. However it seems that reference types in C# are unique when assigned to the instance object. So it is not possible to easily destroy or lose a reference to an instance of an object.