[Text Form - Control creation and events]
Created a couple classes which can create ui buttons that are parented to a control and positioned to fit within the panel.
Rows and columns can be specified to say how it will layout the buttons on the parent panel.

Currently can have the buttons pressed to modify the text box of the form application. Will create a separate project now to create a Form version of my calculator application.