[Calculator Form Application - Migrated Console Application + Text Form Project]
Migrated over the console version of the calculator application and combined it with the functionality from the text form project.

Now has a GUI display that can perform integer operations on an expression built from using the form buttons.

Currently only has support for the console versions list of operations, all others throw an error for now.
Also added a very quick implementation of clearing the text field or appending to it depending on the result displayed after a calculation.